src/aiservice/AIruntime/requirement.txt
google-cloud-aiplatform
mcp
